feat: fetch picking list data on login
